intcode: ALU extension.
Normally changing an exposed type like this would be a breaking change,
but there's an informal contract (between me and... me) not to depend on it.

+++ intcode/src/lib.rs
@@ -6,7 +6,7 @@ pub mod exec;
 
 pub use exec::{Computer, Device, ExecError, IOError, Stepped};
 
-pub type Word = i32;
+pub type Word = i64;
 
 pub type ParseError = ParseIntError;
